refactor(/view, /model): dto 들을 model 에서 dto 패키지로 이동해 레이어 간 분리를 명확히 해라
