+++ src/main/java/racingcar/controller/GameController.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
 import racingcar.model.car.Cars;
-import racingcar.model.car.RoundSnapshot;
-import racingcar.model.car.WinnersDto;
+import racingcar.dto.RoundSnapshot;
+import racingcar.dto.WinnersDto;
 import racingcar.model.game.Game;
 import racingcar.model.game.NumberPicker;
 import racingcar.model.game.RandomNumberPicker;
